+++ discovery/Dockerfile
@@ -3,7 +3,7 @@ FROM openjdk:17-alpine
 
 WORKDIR /null
 
-COPY /build/libs/discovery.jar /null/app.jar
+COPY /build/libs/discovery-0.0.1.jar /null/app.jar
 ENTRYPOINT java -jar app.jar
 
-LABEL version="unspecified"
+LABEL version="0.0.1"
