+++ shop-user-service/src/main/java/com/example/shopuserservice/web/security/ReactiveSecurityConfig.java
@@ -3,6 +3,7 @@ package com.example.shopuserservice.web.security;
 import com.example.shopuserservice.domain.user.data.User;
 import com.example.shopuserservice.domain.user.redisrepository.UserRefreshTokenRedisRepository;
 import com.example.shopuserservice.domain.user.repository.UserRepository;
+import com.example.shopuserservice.web.security.filter.JwtRefreshTokenAuthenticationFilter;
 import com.example.shopuserservice.web.security.filter.JwtTokenAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
