--- shop-user-service/src/main/java/com/example/shopuserservice/web/security/token/TokenInfo.java
@@ -1,14 +1,0 @@
-package com.example.shopuserservice.web.security.token;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-@AllArgsConstructor
-public class TokenInfo {
-    private String grantType;
-    private String accessToken;
-    private String refreshToken;
-}
