+++ spring-chatting-front-server/Dockerfile
@@ -1,9 +1,11 @@
 
 FROM openjdk:17-alpine
 
-WORKDIR /null
+WORKDIR /app
 
-COPY /build/libs/spring-chatting-front-server.jar /null/app.jar
+COPY /build/libs/spring-chatting-front-server-0.0.1.jar /app/app.jar
+# This part should be stored in the secret storage
+COPY keystore.p12 /app/keystore.p12
 ENTRYPOINT java -jar app.jar
 
-LABEL version="unspecified"
+LABEL version="0.0.1"
